Deprecate implicit Route conversions
