+++ core/src/it/scala/fr/davit/akka/http/metrics/core/HttpMetricsItSpec.scala
@@ -82,7 +82,7 @@ class HttpMetricsItSpec
 
     val binding = Http()
       .newMeteredServerAt("localhost", 0, registry)
-      .bindFlow(route)
+      .bindFlow(HttpMetrics.metricsRouteToFlow(route))
       .futureValue
 
     val uri = Uri()
@@ -106,7 +106,7 @@ class HttpMetricsItSpec
 
     val binding = Http()
       .newMeteredServerAt("localhost", 0, registry)
-      .bind(route)
+      .bind(HttpMetrics.metricsRouteToFunction(route))
       .futureValue
 
     val uri = Uri()
@@ -130,7 +130,7 @@ class HttpMetricsItSpec
 
     val binding = Http()
       .newMeteredServerAt("localhost", 0, registry)
-      .bindFlow(route)
+      .bindFlow(HttpMetrics.metricsRouteToFlow(route))
       .futureValue
 
     val uri = Uri()
