+++ dropwizard/src/it/scala/fr/davit/akka/http/metrics/dropwizard/DropwizardMetricsItSpec.scala
@@ -17,7 +17,6 @@
 package fr.davit.akka.http.metrics.dropwizard
 
 import java.util.concurrent.TimeUnit
-
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
@@ -31,6 +30,7 @@ import fr.davit.akka.http.metrics.core.scaladsl.server.HttpMetricsDirectives._
 import fr.davit.akka.http.metrics.dropwizard.marshalling.DropwizardMarshallers._
 import com.codahale.metrics.MetricRegistry
 import com.codahale.metrics.jvm.{CachedThreadStatesGaugeSet, GarbageCollectorMetricSet, MemoryUsageGaugeSet}
+import fr.davit.akka.http.metrics.core.HttpMetrics
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.flatspec.AnyFlatSpecLike
@@ -72,7 +72,7 @@ class DropwizardMetricsItSpec
 
     val binding = Http()
       .newMeteredServerAt("localhost", 0, registry)
-      .bindFlow(route)
+      .bindFlow(HttpMetrics.metricsRouteToFlow(route))
       .futureValue
 
     val uri = Uri("/metrics")
